+++ tests/test_stream.py
@@ -417,13 +417,13 @@ class TestStanzaStream(StanzaStreamTestBase):
 
     @unittest.skipIf(aioxmpp.version_info >= (1, 0, 0),
                      "not applying to this version of aioxmpp")
-    def test_register_iq_request_coro_casts_enum_and_warn(self):
+    def test_register_iq_request_handler_casts_enum_and_warn(self):
         self.stream._ALLOW_ENUM_COERCION = True
         with self.assertWarnsRegex(
                 DeprecationWarning,
                 r"passing a non-enum value as type_ is deprecated and will "
                 "be invalid as of aioxmpp 1.0") as ctx:
-            self.stream.register_iq_request_coro(
+            self.stream.register_iq_request_handler(
                 "get",
                 FancyTestIQ,
                 unittest.mock.sentinel.coro,
@@ -437,31 +437,31 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaisesRegex(
                 ValueError,
                 r"only one listener is allowed per tag"):
-            self.stream.register_iq_request_coro(
+            self.stream.register_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ,
                 unittest.mock.sentinel.coro,
             )
 
-    def test_register_iq_request_coro_raises_on_string_type(self):
+    def test_register_iq_request_handler_raises_on_string_type(self):
         if aioxmpp.version_info < (1, 0, 0):
             self.stream._ALLOW_ENUM_COERCION = False
 
         with self.assertRaisesRegex(
                 TypeError,
                 r"type_ must be IQType, got .*"):
-            self.stream.register_iq_request_coro(
+            self.stream.register_iq_request_handler(
                 "get",
                 FancyTestIQ,
                 unittest.mock.sentinel.coro,
             )
 
-    def test_register_iq_request_coro_does_not_warn_on_enum(self):
+    def test_register_iq_request_handler_does_not_warn_on_enum(self):
         self.stream._ALLOW_ENUM_COERCION = True
 
         with warnings.catch_warnings(record=True) as w:
             warnings.simplefilter("always")
-            self.stream.register_iq_request_coro(
+            self.stream.register_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ,
                 unittest.mock.sentinel.coro,
@@ -469,12 +469,12 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.assertFalse(w)
 
-    def test_register_iq_request_coro_rejects_duplicate_registration(self):
+    def test_register_iq_request_handler_rejects_duplicate_registration(self):
         @asyncio.coroutine
         def handle_request(stanza):
             pass
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -484,15 +484,15 @@ class TestStanzaStream(StanzaStreamTestBase):
             pass
 
         with self.assertRaises(ValueError):
-            self.stream.register_iq_request_coro(
+            self.stream.register_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ,
                 handle_request)
 
     @unittest.skipIf(aioxmpp.version_info >= (1, 0, 0),
                      "not applying to this version of aioxmpp")
-    def test_unregister_iq_request_coro_casts_enum_and_warn(self):
-        self.stream.register_iq_request_coro(
+    def test_unregister_iq_request_handler_casts_enum_and_warn(self):
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             unittest.mock.sentinel.coro,
@@ -503,7 +503,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 DeprecationWarning,
                 r"passing a non-enum value as type_ is deprecated and will "
                 "be invalid as of aioxmpp 1.0") as ctx:
-            self.stream.unregister_iq_request_coro(
+            self.stream.unregister_iq_request_handler(
                 "get",
                 FancyTestIQ,
             )
@@ -514,13 +514,13 @@ class TestStanzaStream(StanzaStreamTestBase):
         )
 
         with self.assertRaises(KeyError):
-            self.stream.unregister_iq_request_coro(
+            self.stream.unregister_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ,
             )
 
-    def test_unregister_iq_request_coro_raises_on_string_type(self):
-        self.stream.register_iq_request_coro(
+    def test_unregister_iq_request_handler_raises_on_string_type(self):
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             unittest.mock.sentinel.coro,
@@ -532,13 +532,13 @@ class TestStanzaStream(StanzaStreamTestBase):
         with self.assertRaisesRegex(
                 TypeError,
                 r"type_ must be IQType, got .*"):
-            self.stream.unregister_iq_request_coro(
+            self.stream.unregister_iq_request_handler(
                 "get",
                 FancyTestIQ,
             )
 
-    def test_unregister_iq_request_coro_does_not_warn_on_enum(self):
-        self.stream.register_iq_request_coro(
+    def test_unregister_iq_request_handler_does_not_warn_on_enum(self):
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             unittest.mock.sentinel.coro,
@@ -548,17 +548,17 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         with warnings.catch_warnings(record=True) as w:
             warnings.simplefilter("always")
-            self.stream.unregister_iq_request_coro(
+            self.stream.unregister_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ,
             )
 
         self.assertFalse(w)
 
-    def test_register_iq_request_coro_raises_on_response_IQType(self):
+    def test_register_iq_request_handler_raises_on_response_IQType(self):
         for member in structs.IQType:
             if member.is_request:
-                self.stream.register_iq_request_coro(
+                self.stream.register_iq_request_handler(
                     member,
                     FancyTestIQ,
                     unittest.mock.sentinel.coro,
@@ -567,7 +567,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 with self.assertRaisesRegex(
                         ValueError,
                         r".* is not a request IQType"):
-                    self.stream.register_iq_request_coro(
+                    self.stream.register_iq_request_handler(
                         member,
                         FancyTestIQ,
                         unittest.mock.sentinel.coro,
@@ -585,7 +585,7 @@ class TestStanzaStream(StanzaStreamTestBase):
             response_payload = FancyTestIQ()
             return response_payload
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -614,7 +614,7 @@ class TestStanzaStream(StanzaStreamTestBase):
             fut.set_result(response_payload)
             return fut
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -659,7 +659,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         def handle_request(stanza):
             raise Exception("foo")
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -687,7 +687,7 @@ class TestStanzaStream(StanzaStreamTestBase):
         def handle_request(stanza):
             raise Exception("foo")
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -719,7 +719,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 text="foobarbaz",
             )
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -754,7 +754,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 text="foobarbaz",
             )
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
@@ -777,13 +777,13 @@ class TestStanzaStream(StanzaStreamTestBase):
 
         self.stream.stop()
 
-    def test_unregister_iq_request_coro_raises_if_none_was_registered(self):
+    def test_unregister_iq_request_handler_raises_if_none_was_registered(self):
         with self.assertRaises(KeyError):
-            self.stream.unregister_iq_request_coro(
+            self.stream.unregister_iq_request_handler(
                 structs.IQType.GET,
                 FancyTestIQ)
 
-    def test_unregister_iq_request_coro(self):
+    def test_unregister_iq_request_handler(self):
         iq = make_test_iq()
         iq.autoset_id()
 
@@ -794,11 +794,11 @@ class TestStanzaStream(StanzaStreamTestBase):
             nonlocal recvd
             recvd = stanza
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             handle_request)
-        self.stream.unregister_iq_request_coro(
+        self.stream.unregister_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ)
 
@@ -1491,7 +1491,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 exc = inner_exc
                 raise
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             coro,
@@ -1525,7 +1525,7 @@ class TestStanzaStream(StanzaStreamTestBase):
                 exc = inner_exc
                 raise
 
-        self.stream.register_iq_request_coro(
+        self.stream.register_iq_request_handler(
             structs.IQType.GET,
             FancyTestIQ,
             coro,
@@ -3580,6 +3580,66 @@ class TestStanzaStream(StanzaStreamTestBase):
             result,
             reply.payload,
         )
+
+    @unittest.skipIf(aioxmpp.version_info >= (1, 0, 0),
+                     "not applying to this version of aioxmpp")
+    def test_register_iq_request_coro_warns_and_forwards_to_handler(self):
+        with contextlib.ExitStack() as stack:
+            handler = stack.enter_context(unittest.mock.patch.object(
+                self.stream,
+                "register_iq_request_handler",
+            ))
+
+            stack.enter_context(
+                self.assertWarnsRegex(
+                    DeprecationWarning,
+                    r"register_iq_request_coro is a deprecated alias to "
+                    r"register_iq_request_handler and will be removed in "
+                    r"aioxmpp 1.0")
+            )
+
+            result = self.stream.register_iq_request_coro(
+                unittest.mock.sentinel.type_,
+                unittest.mock.sentinel.payload_class,
+                unittest.mock.sentinel.coro,
+            )
+
+        handler.assert_called_once_with(
+            unittest.mock.sentinel.type_,
+            unittest.mock.sentinel.payload_class,
+            unittest.mock.sentinel.coro,
+        )
+
+        self.assertEqual(result, handler())
+
+    @unittest.skipIf(aioxmpp.version_info >= (1, 0, 0),
+                     "not applying to this version of aioxmpp")
+    def test_unregister_iq_request_coro_warns_and_forwards_to_handler(self):
+        with contextlib.ExitStack() as stack:
+            handler = stack.enter_context(unittest.mock.patch.object(
+                self.stream,
+                "unregister_iq_request_handler",
+            ))
+
+            stack.enter_context(
+                self.assertWarnsRegex(
+                    DeprecationWarning,
+                    r"unregister_iq_request_coro is a deprecated alias to "
+                    r"unregister_iq_request_handler and will be removed in "
+                    r"aioxmpp 1.0")
+            )
+
+            result = self.stream.unregister_iq_request_coro(
+                unittest.mock.sentinel.type_,
+                unittest.mock.sentinel.payload_class,
+            )
+
+        handler.assert_called_once_with(
+            unittest.mock.sentinel.type_,
+            unittest.mock.sentinel.payload_class,
+        )
+
+        self.assertEqual(result, handler())
 
 
 class TestStanzaStreamSM(StanzaStreamTestBase):
@@ -5650,7 +5710,7 @@ class Testiq_handler(unittest.TestCase):
     def test_enter_registers_coroutine(self):
         self.cm.__enter__()
 
-        self.stream.register_iq_request_coro.assert_called_with(
+        self.stream.register_iq_request_handler.assert_called_with(
             unittest.mock.sentinel.iqtype,
             unittest.mock.sentinel.payload,
             unittest.mock.sentinel.coro,
@@ -5662,7 +5722,7 @@ class Testiq_handler(unittest.TestCase):
 
         self.cm.__exit__(None, None, None)
 
-        self.stream.unregister_iq_request_coro.assert_called_with(
+        self.stream.unregister_iq_request_handler.assert_called_with(
             unittest.mock.sentinel.iqtype,
             unittest.mock.sentinel.payload,
         )
@@ -5679,7 +5739,7 @@ class Testiq_handler(unittest.TestCase):
 
         result = self.cm.__exit__(*info)
 
-        self.stream.unregister_iq_request_coro.assert_called_with(
+        self.stream.unregister_iq_request_handler.assert_called_with(
             unittest.mock.sentinel.iqtype,
             unittest.mock.sentinel.payload,
         )